Allow user to override the trusted DD AWS account id

+++ aws/release.sh
@@ -12,7 +12,20 @@ else
     BUCKET=$1
 fi
 
+# Confirm to proceed
+for i in *.yaml; do
+    [ -f "$i" ] || break
+    echo "About to upload $i to s3://datadog-cloudformation-template/aws/$i"
+done
+read -p "Continue (y/n)?" CONT
+if [ "$CONT" != "y" ]; then
+  echo "Exiting"
+  exit 1
+fi
+
 aws s3 cp . s3://${BUCKET}/aws --recursive --exclude "*" --include "*.yaml" \
     --grants read=uri=http://acs.amazonaws.com/groups/global/AllUsers
+echo "Done uploading the template, and here is the CloudFormation quick launch URL"
+echo "https://console.aws.amazon.com/cloudformation/home#/stacks/create/review?stackName=datadog-aws-integration&templateURL=https://${BUCKET}.s3.amazonaws.com/aws/main.yaml"
 
 echo "Done!"
